无

+++ src/main/java/net/hhdsj/goodblock/procedures/Paletreespa1DangPiLinFangKuaiGengXinShiProcedure.java
@@ -6,7 +6,7 @@ import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.core.BlockPos;
 
 public class Paletreespa1DangPiLinFangKuaiGengXinShiProcedure {
-	public static void execute(LevelAccessor world, double x, double y, double z) {
+	public static void execute(LevelAccessor world, int x, int y, int z) {
 		if ((world.getBlockState(new BlockPos(x, y - 1, z))) == Blocks.AIR.defaultBlockState()) {
 			{
 				BlockPos _pos = new BlockPos(x, y, z);
